initial addition of the mkfs_vfat command
Currently a very simply wrapper for the mkfs.vfat command that
allows for more FAT specific options to be passed through than the current
mkfs command handles.

The focus with this commit was to become familiar with adding
additional RPC actions to the daemon. With that complete, the next
step will be adding more complete support for all the mkfs.vfat options.

Purpose:
libguestfs is obviously wonderful for generating disk images for virtual
machines. It is also wonderful for generating disk images intended for use
with various SBC (Single Board Computers) like the Raspberry Pi or BeagleBone.

It is often the case that the bootloader for these devices can only boot
specifically configured FAT filesystems and the options available to mkfs
are simply not rich enough to cover this use case. It is the purpose of this
command to fully wrap mkfs.vfat and passthrough any option an image might need
to be bootable by such devices.

Paired with the rest of the rich libguestfs API, this command will make image
creation for SBCs much less frustrating.